+++ src/Translation/LocaleHelper.php
@@ -1,0 +1,14 @@
+<?php declare(strict_types=1);
+
+namespace Torr\Storyblok\Translation;
+
+final class LocaleHelper
+{
+	/**
+	 * Returns whether the given string is a syntactically valid locale.
+	 */
+	public static function isValidLocale (string $value) : bool
+	{
+		return 1 === \preg_match('~^\\w+(-\\w+)?$~', $value);
+	}
+}